register survey component removed and combined with register component fixed other issues
