--- src/components/RegisterSurveyComponent/index.js
@@ -1,2 +1,0 @@
-import RegisterSurveyComponent from "./RegisterSurveyComponent";
-export default RegisterSurveyComponent;
