+++ src/pages/AuthorizationPage/AuthorizationPage.module.scss
@@ -1,5 +1,6 @@
 .AuthorizationPage {
   height: 100%;
+  margin: 0;
   width: 100%;
   display: flex;
   flex-direction: row;
